+++ lib/boruta/adapters/ecto/admin.ex
@@ -9,6 +9,8 @@ defmodule Boruta.Ecto.Admin do
   defdelegate update_client(client, attrs), to: Boruta.Ecto.Admin.Clients
   defdelegate regenerate_client_secret(client), to: Boruta.Ecto.Admin.Clients
   defdelegate regenerate_client_secret(client, secret), to: Boruta.Ecto.Admin.Clients
+  defdelegate regenerate_client_key_pair(client), to: Boruta.Ecto.Admin.Clients
+  defdelegate regenerate_client_key_pair(client, public_key, private_key), to: Boruta.Ecto.Admin.Clients
   defdelegate delete_client(client), to: Boruta.Ecto.Admin.Clients
 
   defdelegate list_scopes, to: Boruta.Ecto.Admin.Scopes